feat: add default repo config for argo
