+++ internal/pkg/plugin/argocd/create.go
@@ -11,11 +11,13 @@ import (
 
 // Create creates ArgoCD with provided options.
 func Create(options map[string]interface{}) (map[string]interface{}, error) {
-	// 1. decode options
+	// 1. decode options and fill default options if miss
 	var opts Options
 	if err := mapstructure.Decode(options, &opts); err != nil {
 		return nil, err
 	}
+
+	defaultMissedOptions(&opts)
 
 	if errs := validate(&opts); len(errs) != 0 {
 		for _, e := range errs {
